api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
